#108 Energy decreases by 1 on choosing wrong response. Changed type of energyInit to int. Increase and decrease is now in int.

+++ lib/widgets/energy_bar.dart
@@ -7,7 +7,7 @@ class EnergyBar extends StatefulWidget{
 }
 
 class _EnergyBarState extends State<EnergyBar>{
-  double _energy;
+  int _energy;
 
   @override
   void initState(){
